Sort: don't allocate sort struct on heap
Wow, this was fun to find.  It looks like anything passed to sort.Sort is thought of as escaping to the heap - https://github.com/golang/go/issues/15451.
There's an easy fix for this which is to allocate the struct in global context; since everything is single-threaded no issues with this.  I looked a little bit at using sort.Slice(); we can't use this because we want to sort one array based on the other.  After the first swaps happen in the original array the moveScore array gets out of sync.

This was a leftover from using a priority queue for the sorting interface.